pyproject.toml
[tool.poetry]
name = "vision-test"
version = "0.1.0"
description = ""
authors = ["Mikhail Drozdov <deathlokmike@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.2"
asyncpg = "^0.29.0"
psutil = "^5.9.6"
websockets = "^12.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
nuitka = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
